@@ -2,30 +2,35 @@
 
 namespace Main
 {
-    class HomeZone : MonoBehaviour
+    internal class HomeZone : MonoBehaviour
     {
-        public ETeam Team = ETeam.A;
+        public ETeam Team { set; get; } = ETeam.A;
 
         private float m_SqrRadius;
-        void Start()
+
+        private void Start()
         {
             m_SqrRadius = Match.instance.GlobalSetting.HomeZoneRadius * Match.instance.GlobalSetting.HomeZoneRadius;
         }
-        void Update()
+
+        private void Update()
         {
-            Tank t = Match.instance.GetTank(Team);
-            if(t == null)
+            var tanks = Match.instance.GetTanks(Team);
+            if(tanks == null)
             {
                 return;
             }
-            Vector3 homeZonePos = Match.instance.GetRebornPos(Team);
-            if((homeZonePos - t.Position).sqrMagnitude < m_SqrRadius)
+            foreach (var t in tanks)
             {
-                t.HPRecovery(Time.deltaTime * Match.instance.GlobalSetting.HPRecoverySpeed);
-            }
-            else
-            {
-                t.HPRecovery(0);
+                Vector3 homeZonePos = Match.instance.GetRebornPos(Team);
+                if((homeZonePos - t.Position).sqrMagnitude < m_SqrRadius)
+                {
+                    t.HPRecovery(Time.deltaTime * Match.instance.GlobalSetting.HPRecoverySpeed);
+                }
+                else
+                {
+                    t.HPRecovery(0);
+                }
             }
         }
     }
